requirements.txt
akshare>=1.14.0
pandas>=2.0.0
numpy>=1.24.0
tabulate>=0.9.0
pyyaml>=6.0
